Added GCompiler::compileModule
Assembly member of GCompiler is now a pointer to the currently being compiled Assembly. Since the compiler (currently) only needs access to the Assembly while compiling there is no need to store the full Assembly in the compiler. This also lets us return the Assembly directly after compiling without any moving or copying